% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{bd09togcj02}
\alias{bd09togcj02}
\title{From bd09 to gcj02.}
\usage{
bd09togcj02(bd_lon, bd_lat)
}
\arguments{
\item{bd_lon}{A number.}

\item{bd_lat}{A number.}
}
\value{
result.
}
\description{
From bd09 to gcj02.
}
