% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{gcj02towgs84}
\alias{gcj02towgs84}
\title{From gcj02 to wgs84}
\usage{
gcj02towgs84(lng, lat)
}
\arguments{
\item{lng}{A number.}

\item{lat}{A number.}
}
\value{
result.
}
\description{
From gcj02 to wgs84
}
